Import-helper-function for excel-sheets

Note that the package comes with a template-xlsx-file that contains all the columns that are expected by the package to work properly. This template is located in the template-folder of the installed package. Alternatively, you can download the template from https://github.com/MarselScheer/projectPlan/inst/template

import_xlsx(fName, sheets)

Arguments

fName
File name of the xlsx-file that should be imported

sheets
list of sheet-names that are imported. If not specified all sheets from the xlsx-file are imported. If the sheet does not contain a project-column, then the name of the sheet is used as a project name.

Value

One data.table containing all imported sheets. Usually, such an object needs to be preprocessed by wrangle_raw_plan

See also

Examples

import_xlsx(system.file("template","projects.xlsx", package = "projectPlan"))

#>           section        id          depends_on    start  end deadline
#> 1:         0_prep      <NA>                <NA>    TODAY <NA>     <NA>
#> 2:         0_prep roughplan            PREVIOUS     <NA> <NA>     <NA>
#> 3:         0_prep      <NA>                <NA> PREVIOUS <NA>     <NA>
#> 4:        1_offer      <NA> PREVIOUS, roughplan     <NA> <NA>     <NA>
#> 5:        1_offer      plan            PREVIOUS     <NA> <NA>     <NA>
#> 6:        1_offer       buy            PREVIOUS     <NA> <NA>     <NA>
#> 7: 2_installation      <NA>                plan     <NA> <NA>     <NA>
#> 8: 2_installation      <NA>       PREVIOUS, buy     <NA> <NA>    43868
#>    est_duration      status   resource
#> 1:            1 unscheduled          I
#> 2:            2 unscheduled          I
#> 3:            2 unscheduled  I, F1, F2
#> 4:            5 unscheduled I, Company
#> 5:            1 unscheduled I, Company
#> 6:            1 unscheduled          I
#> 7:            1 unscheduled          I
#> 8:            1 unscheduled I, Company
#>                                              task progress
#> 1:                                measure kitchen        0
#> 2:                              make a rough plan        0
#> 3: ask friends which companies they can recommend        0
#> 4:     make appointment with company for planning        0
#> 5:                               plan new kitchen        0
#> 6:                                buy new kitchen        0
#> 7:              make appointment for installation        0
#> 8:                installation of the new kitchen        0
#>                                                                                                                            comments
#> 1:                                                                                                                             <NA>
#> 2:                                                                                                                             <NA>
#> 3: Yyyy-mm-dd: \n.ask-for-XYZ: The company XYZ has opened a new store here. Should try to find if someone has experience with them.
#> 4:                                                                                                                             <NA>
#> 5:                                                                                                                             <NA>
#> 6:                                                                                                                             <NA>
#> 7:                                                                                                                             <NA>
#> 8:                                                                                                                             <NA>
#>                                                  microtasks  now project
#> 1:                                                     <NA> <NA> kitchen
#> 2:                                                     <NA> <NA> kitchen
#> 3:                                             .ask-for-XYZ <NA> kitchen
#> 4:                                                     <NA> <NA> kitchen
#> 5:                                                     <NA> <NA> kitchen
#> 6:                                                     <NA> <NA> kitchen
#> 7:                                .get-written-confirmation <NA> kitchen
#> 8: .check-fridge\n.check-oven\n.check-stove\n.check-drawers <NA> kitchen